add support for OS X and iOS as host system
Conflicts:
    README.md
    cctools/autogen.sh
    cctools/configure.ac
    cctools/ld64/src/3rd/helper.c
    cctools/ld64/src/other/Makefile.am

+++ cctools/ld64/src/ld/passes/branch_island.cpp
@@ -498,7 +498,7 @@ void doPass(const Options& opts, ld::Internal& state)
 				if ( target->section().type() == ld::Section::typeStub )
 					dstAddr = totalTextSize;
 				int64_t displacement = dstAddr - srcAddr;
-				TargetAndOffset finalTargetAndOffset = { target, addend };
+				TargetAndOffset finalTargetAndOffset = { target, static_cast<uint32_t>(addend) };
 				const int64_t kBranchLimit = kBetweenRegions;
 				if ( displacement > kBranchLimit ) {
 					// create forward branch chain
